requirements.txt
blinker==1.7.0
click==8.1.7
exceptiongroup==1.1.3
Flask==3.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
PyJWT==2.8.0
PyMySQL==1.1.0
pytest==7.4.3
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
tomli==2.0.1
Werkzeug==3.0.1
